refactor pm quality policy detail router
