+++ src/app/router_registry.py
@@ -165,6 +165,9 @@ from app.routers.dpm_command_center_pm_quality_policies import (
 from app.routers.dpm_command_center_pm_quality_policy_actions import (
     router as dpm_command_center_pm_quality_policy_actions_router,
 )
+from app.routers.dpm_command_center_pm_quality_policy_detail import (
+    router as dpm_command_center_pm_quality_policy_detail_router,
+)
 from app.routers.dpm_command_center_pm_quality_review_action_detail import (
     router as dpm_command_center_pm_quality_review_action_detail_router,
 )
@@ -553,6 +556,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_pm_quality_fairness_router)
     app.include_router(dpm_command_center_pm_quality_fairness_detail_router)
     app.include_router(dpm_command_center_pm_quality_fairness_lookup_router)
+    app.include_router(dpm_command_center_pm_quality_policy_detail_router)
     app.include_router(dpm_command_center_pm_quality_policies_router)
     app.include_router(dpm_command_center_pm_quality_policy_actions_router)
     app.include_router(dpm_command_center_pm_quality_review_action_preview_router)
