+++ src/app/routers/dpm_command_center_pm_quality_policies.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Path, Query
+from fastapi import APIRouter, Query
 
 from app.contracts.dpm_command_center import DpmPmOperatingQualityGatewayResponse
 from app.middleware.correlation import correlation_id_var
@@ -39,34 +39,5 @@ async def list_pm_operating_quality_policies(
             "limit": limit,
             "offset": offset,
         },
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/pm-operating-quality/policies/{policy_id}/versions/{policy_version}",
-    response_model=DpmPmOperatingQualityGatewayResponse,
-    summary="Get PM operating quality policy",
-    description=(
-        "What: returns one persisted manage-owned PM operating quality policy version. When: use "
-        "for audit and score-run preparation. How: Gateway retrieves immutable Manage policy "
-        "configuration without computing or approving PM scores locally."
-    ),
-)
-async def get_pm_operating_quality_policy(
-    policy_id: str = Path(
-        ...,
-        description="Manage-owned PM operating quality policy identifier.",
-        examples=["pmq_sg_dpm"],
-    ),
-    policy_version: str = Path(
-        ...,
-        description="Manage-owned PM operating quality policy version.",
-        examples=["2026.05"],
-    ),
-) -> DpmPmOperatingQualityGatewayResponse:
-    return await dpm_command_center_service().get_pm_operating_quality_policy(
-        policy_id=policy_id,
-        policy_version=policy_version,
         correlation_id=correlation_id_var.get(),
     )
