+++ src/app/routers/dpm_command_center_pm_quality_policy_detail.py
@@ -1,0 +1,43 @@
+from fastapi import APIRouter, Path
+
+from app.contracts.dpm_command_center import DpmPmOperatingQualityGatewayResponse
+from app.middleware.correlation import correlation_id_var
+from app.routers.dpm_command_center_pm_quality_common import (
+    UPSTREAM_PM_OPERATING_QUALITY_ERROR_RESPONSES,
+)
+from app.services.dpm_service_provider import dpm_command_center_service
+
+router = APIRouter(
+    prefix="/api/v1/dpm/command-center",
+    tags=["DPM Command Center"],
+    responses=UPSTREAM_PM_OPERATING_QUALITY_ERROR_RESPONSES,
+)
+
+
+@router.get(
+    "/pm-operating-quality/policies/{policy_id}/versions/{policy_version}",
+    response_model=DpmPmOperatingQualityGatewayResponse,
+    summary="Get PM operating quality policy",
+    description=(
+        "What: returns one persisted manage-owned PM operating quality policy version. When: use "
+        "for audit and score-run preparation. How: Gateway retrieves immutable Manage policy "
+        "configuration without computing or approving PM scores locally."
+    ),
+)
+async def get_pm_operating_quality_policy(
+    policy_id: str = Path(
+        ...,
+        description="Manage-owned PM operating quality policy identifier.",
+        examples=["pmq_sg_dpm"],
+    ),
+    policy_version: str = Path(
+        ...,
+        description="Manage-owned PM operating quality policy version.",
+        examples=["2026.05"],
+    ),
+) -> DpmPmOperatingQualityGatewayResponse:
+    return await dpm_command_center_service().get_pm_operating_quality_policy(
+        policy_id=policy_id,
+        policy_version=policy_version,
+        correlation_id=correlation_id_var.get(),
+    )
